pipeline/pilot_expeiments/PILOT_EXPERIMENT_2/requirements-working.txt
accelerate==0.29.3
action-msgs==1.2.2
action-tutorials-interfaces==0.20.6
action-tutorials-py==0.20.6
actionlib-msgs==4.9.0
ament-cmake-test==1.3.12
ament-copyright==0.12.14
ament-cppcheck==0.12.14
ament-cpplint==0.12.14
ament-flake8==0.12.14
ament-index-python==1.4.0
ament-lint==0.12.14
ament-lint-cmake==0.12.14
ament-package==0.14.1
ament-pep257==0.12.14
ament-uncrustify==0.12.14
ament-xmllint==0.12.14
angles==1.15.0
astra-camera-msgs==1.0.1
bitsandbytes==0.42.0
bond==4.1.2
builtin-interfaces==1.2.2
certifi==2022.12.7
charset-normalizer==2.1.1
composition-interfaces==1.2.2
costmap-converter-msgs==0.1.2
cv-bridge==3.2.1
demo-nodes-py==0.20.6
diagnostic-msgs==4.9.0
diagnostic-updater==4.0.6
domain-coordinator==0.10.0
draccus==0.8.0
dwb-msgs==1.1.19
einops==0.8.2
example-interfaces==0.9.3
examples-rclpy-executors==0.15.4
examples-rclpy-minimal-action-client==0.15.4
examples-rclpy-minimal-action-server==0.15.4
examples-rclpy-minimal-client==0.15.4
examples-rclpy-minimal-publisher==0.15.4
examples-rclpy-minimal-service==0.15.4
examples-rclpy-minimal-subscriber==0.15.4
filelock==3.25.2
fsspec==2026.2.0
gazebo-msgs==3.9.0
geometry-msgs==4.9.0
hf-xet==1.4.3
huggingface_hub==0.36.2
idna==3.4
image-geometry==3.2.1
interactive-markers==2.3.2
Jinja2==3.1.6
laser-geometry==2.4.1
launch==1.0.11
launch-ros==0.19.12
launch-testing==1.0.11
launch-testing-ros==0.19.12
launch-xml==1.0.11
launch-yaml==1.0.11
lifecycle-msgs==1.2.2
logging-demo==0.20.6
map-msgs==2.1.0
MarkupSafe==3.0.3
mergedeep==1.3.4
message-filters==4.3.11
mpmath==1.3.0
mypy_extensions==1.1.0
nav-2d-msgs==1.1.19
nav-msgs==4.9.0
nav2-common==1.1.19
nav2-msgs==1.1.19
nav2-simple-commander==1.0.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.9.86
nvidia-nvtx-cu12==12.1.105
-e git+https://github.com/openvla/openvla.git@c8f03f48af692657d3060c19588038c7220e9af9#egg=openvla
osrf-pycommon==2.1.6
packaging==26.0
pcl-msgs==1.0.0
pendulum-msgs==0.20.6
pillow==12.1.1
psutil==7.2.2
python-qt-binding==1.1.3
PyYAML==6.0.3
pyyaml-include==1.4.1
qt-dotgraph==2.2.4
qt-gui==2.2.4
qt-gui-cpp==2.2.4
qt-gui-py-common==2.2.4
quality-of-service-demo-py==0.20.6
rcl-interfaces==1.2.2
rclpy==3.3.17
rcutils==5.1.7
regex==2026.4.4
requests==2.28.1
resource-retriever==3.1.3
rmw-dds-common==1.6.0
ros2action==0.18.14
ros2bag==0.15.15
ros2cli==0.18.14
ros2component==0.18.14
ros2doctor==0.18.14
ros2interface==0.18.14
ros2launch==0.19.12
ros2lifecycle==0.18.14
ros2multicast==0.18.14
ros2node==0.18.14
ros2param==0.18.14
ros2pkg==0.18.14
ros2run==0.18.14
ros2service==0.18.14
ros2topic==0.18.14
rosbag2-interfaces==0.15.15
rosbag2-py==0.15.15
# Editable install with no version control (rosboard==1.2.1)
-e /home/ros2/yahboomcar_ws/install/rosboard/lib/python3.10/site-packages
rosbridge-msgs==1.3.1
rosgraph-msgs==1.2.2
rosidl-adapter==3.1.7
rosidl-cli==3.1.7
rosidl-cmake==3.1.7
rosidl-generator-c==3.1.7
rosidl-generator-cpp==3.1.7
rosidl-generator-py==0.14.6
rosidl-parser==3.1.7
rosidl-runtime-py==0.9.3
rosidl-typesupport-c==2.0.2
rosidl-typesupport-cpp==2.0.2
rosidl-typesupport-fastrtps-c==2.2.3
rosidl-typesupport-fastrtps-cpp==2.2.3
rosidl-typesupport-introspection-c==3.1.7
rosidl-typesupport-introspection-cpp==3.1.7
rpyutils==0.2.2
rqt-action==2.0.1
rqt-bag==1.1.5
rqt-bag-plugins==1.1.5
rqt-console==2.0.3
rqt-graph==1.3.1
rqt-gui==1.1.8
rqt-gui-py==1.1.8
rqt-msg==1.2.0
rqt-plot==1.1.5
rqt-publisher==1.5.0
rqt-py-common==1.1.8
rqt-py-console==1.0.2
rqt-reconfigure==1.1.2
rqt-service-caller==1.0.5
rqt-shell==1.0.2
rqt-srv==1.0.3
rqt-topic==1.5.0
safetensors==0.7.0
scripts==3.9.0
sensor-msgs==4.9.0
sensor-msgs-py==4.9.0
sentencepiece==0.1.99
shape-msgs==4.9.0
slam-toolbox==2.6.10
smclib==4.1.2
sros2==0.10.7
statistics-msgs==1.2.2
std-msgs==4.9.0
std-srvs==4.9.0
stereo-msgs==4.9.0
sympy==1.14.0
teleop-twist-keyboard==2.4.1
tf2-geometry-msgs==0.25.16
tf2-kdl==0.25.16
tf2-msgs==0.25.16
tf2-py==0.25.16
tf2-ros-py==0.25.16
tf2-sensor-msgs==0.25.16
tf2-tools==0.25.16
timm==0.9.10
tokenizers==0.19.1
toml==0.10.2
topic-monitor==0.20.6
torch==2.2.0+cu121
torchaudio==2.2.0+cu121
torchvision==0.17.0+cu121
tqdm==4.67.3
trajectory-msgs==4.9.0
transformers==4.40.1
triton==2.2.0
turtlesim==1.4.3
typing-inspect==0.9.0
typing_extensions==4.15.0
unique-identifier-msgs==2.2.1
urllib3==1.26.13
visualization-msgs==4.9.0
